Maantenimiento Camara y otros warnings

+++ OpenGL/src/Engine/Game/CamaraOrtografica.hpp
@@ -7,14 +7,14 @@ namespace GL
 	class CamaraOrtografica
 	{
 	public:
-		CamaraOrtografica(uint32_t ancho, uint32_t alto, glm::vec3 posicion);
+		CamaraOrtografica(float ancho, float alto, glm::vec3 posicion);
 		virtual ~CamaraOrtografica();
 
 		CamaraOrtografica(const CamaraOrtografica& other) = delete;
 		CamaraOrtografica& operator==(const CamaraOrtografica& other) = delete;
 
 		void Actualizar();
-		void CambioDimensiones(uint32_t ancho, uint32_t alto);
+		void CambioDimensiones(float ancho, float alto);
 
 		inline glm::mat4 GetMatrixVistaProyeccion() const { return m_MatrizProyeccion * m_MatrizVista; }
 
@@ -22,8 +22,8 @@ namespace GL
 		glm::mat4 m_MatrizVista{ 1.0f };
 		glm::mat4 m_MatrizProyeccion{ 1.0f };
 
-		uint32_t m_Ancho{};
-		uint32_t m_Alto{};
+		float m_Ancho{};
+		float m_Alto{};
 		glm::vec3 m_Arriba{};
 		glm::vec3 m_Posicion{};
 		glm::vec3 m_Centro{};
